Move `getbytes` to internal
