+++ internal/getbytes/getbytes.go
