+++ internal/getbytes/getbytes_test.go
